Rename type-to-config to block-configs
