+++ lib/block-configs.js
