+++ lib/board.js
@@ -1,6 +1,6 @@
 const Block = require('./block');
 const Data = require('./data');
-const typeToConfig = require('./type-to-config');
+const blockConfigs = require('./block-configs');
 const EventEmitter = require('events');
 
 class Board extends EventEmitter {
@@ -97,7 +97,7 @@ class Board extends EventEmitter {
 	}
 
 	placeBlock({x, y, type, rotate}) {
-		const config = Object.assign({}, typeToConfig[type]);
+		const config = Object.assign({}, blockConfigs[type]);
 		config.x = x;
 		config.y = y;
 		config.rotate = rotate;
