+++ lib/validator.js
@@ -1,9 +1,9 @@
-const typeToConfig = require('./type-to-config');
+const blockConfigs = require('./block-configs');
 const stages = require('../stages');
 const Board = require('./board');
 const {arrayEquals, zip} = require('./util');
 
-const blockNames = Object.keys(typeToConfig);
+const blockNames = Object.keys(blockConfigs);
 
 // https://github.com/tsg-ut/mnemo/wiki/%E3%82%B9%E3%82%B3%E3%82%A2%E3%81%AE%E4%BB%95%E6%A7%98#%E8%A8%88%E7%AE%97%E5%BC%8F
 module.exports.calculateScore = ({clocks: C, blocks: B, stage}) => {
